slight bugfix

+++ auroracam/archAndFree.py
@@ -1,7 +1,13 @@
 # copyright (c) Mark McIntyre, 2024-
 #
 # Python script to free diskspace on Auroracam
-#
+# 
+# The script loops through the folders and files in $DATADIR, checking if they're in 
+# FILES_TO_UPLOAD.inf. 
+# If present, the folder is compressed and uploaded to my server.
+# For each folder and file a check is made if there's sufficient free space for the next 
+# night's run and folder or file is deleted if necessary. 
+# At least three days data are always kept. 
 
 import shutil
 import os
@@ -35,11 +41,12 @@ def getFreeSpace():
 
 def getNeededSpace():
     #each jpg is about 100kB, and we capture about 20,000 per day - about one every 4 seconds
-    # plus extra 100MB for the timelapses and tarballs, and a bit of overhead
-    mp4space = 100 * 1024
-    tarballspace = 1500 * 1024
-    extraspace = 50 * 1024
-    reqspace = 20000 * 100 + extraspace + tarballspace + mp4space
+    # plus extra for the timelapses and tarballs, and a bit of overhead
+    jpgspace = 20000 * 100 # 100 kB per file
+    mp4space = 100 * 1024  # 100 MB
+    tarballspace = 1500 * 1024 # 1.5 GB 
+    extraspace = 50 * 1024 # 50 MB extra just in case
+    reqspace = jpgspace + extraspace + tarballspace + mp4space
     return reqspace
 
 
@@ -81,10 +88,10 @@ def compressAndUpload(datadir, thisfile, archserver, archfldr):
                 filestat = ftp_client.stat(uploadfile)
                 log.info(f'uploaded {filestat.st_size} bytes')
                 os.remove(zfname + '.zip')
-                shutil.rmtree(zfname)
             except Exception as e:
                 log.error(f'unable to upload {thisfile}')
                 log.info(e, exc_info=True)
+                return False
         except Exception as e:
             log.error(f'unable to upload {thisfile}')
             log.info(e, exc_info=True)
@@ -113,14 +120,20 @@ def freeUpSpace(datadir, archserver, archfldr):
                 if compressAndUpload(datadir, thisfile, archserver, archfldr):
                     filestoupload.remove(patt)
                     saveFilesToUpload(datadir, filestoupload)
+                    newfree = getFreeSpace()
+                    if newfree < reqkb:
+                        shutil.rmtree(os.path.join(datadir, thisfile))
                 else:
                     print(f'problem compressing or uploading {thisfile}')
+        newfree = getFreeSpace()
         if os.path.isfile(os.path.join(datadir, thisfile)):
-            log.info(f'deleting {thisfile}')
-            os.remove(os.path.join(datadir, thisfile))
+            if newfree < reqkb:
+                log.info(f'deleting {thisfile}')
+                os.remove(os.path.join(datadir, thisfile))
         else:
-            log.info(f'removing folder {thisfile}')
-            shutil.rmtree(os.path.join(datadir, thisfile))
+            if newfree < reqkb:
+                log.info(f'removing folder {thisfile}')
+                shutil.rmtree(os.path.join(datadir, thisfile))
     newfree = getFreeSpace()
     log.info(f'freed {(newfree - freekb)/1024/1024:.3f} GB, have {newfree/1024/1024:.3f} GB')
     return
